+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+* [#69] Hidden Search Bar and Network Selection
 * Hidden Validator Votes
 
 ## v0.1.2