+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [#69] Hidden Search Bar and Network Selection
 * [#68] Added transaction contract input parameters in TransactionDetails
+* Hidden Validator Votes
 
 ## v0.1.2